+++ strawberry/src/core/design_system/foundation/border.tsx
