+++ strawberry/src/core/design_system/index.ts
@@ -1,0 +1,9 @@
+import theme from "./theme";
+
+import Label from "./styles/Label";
+import Line from "./styles/Line";
+import Wrapper from "./styles/Wrapper";
+
+import EventButton from "./buttons/EventButton";
+
+export { theme, Label, Line, Wrapper, EventButton };
