+++ strawberry/src/core/design_system/styles/Label.tsx
