+++ strawberry/src/core/design_system/styles/Line.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import theme from "./theme";
+import theme from "../theme";
 
 interface LineProps {
   $backgroundcolor?: string;
